Moving test_ssh_ldap and test_ssh_ad in test_012_directory_service_ssh
Those 2 test fails on HA
